+++ src/main/scala/org/example/monad/free/Lifting.scala
@@ -3,8 +3,8 @@ package org.example.monad.free
 import org.example.monad.free.Free._
 
 object Lifting {
-  def forward(pos: Position, l: Int): Free[Instruction, Position] = Free.liftF(Forward(pos, l))
-  def backward(pos: Position, l: Int): Free[Instruction, Position] = Free.liftF(Backward(pos, l))
+  def forward(pos: Position, l: Double): Free[Instruction, Position] = Free.liftF(Forward(pos, l))
+  def backward(pos: Position, l: Double): Free[Instruction, Position] = Free.liftF(Backward(pos, l))
   def left(pos: Position, degree: Degree): Free[Instruction, Position] = Free.liftF(RotateLeft(pos, degree))
   def right(pos: Position, degree: Degree): Free[Instruction, Position] = Free.liftF(RotateRight(pos, degree))
   def showPosition(pos: Position): Free[Instruction, Unit] = Free.liftF(ShowPosition(pos))
